Remove quit attempts to exit editor
Before when trying to exit the editor with the buffer modified we would
ask the user to repeat the quit command several times. This is not very
friendly and requires using some global state. Instead we now prompt the
user if they're sure they want to quit, and act accordingly.