Add tab command
Instead of treating a tab as a printable character, which was causing
problems for the user input in the prompt, we don't treat a tab as
printable and add a command for inserting a tab. In the future this
command could be supported to insert a variable number of spaces based
on a particular option.