Remove command initialization
Instead of looking up a command using the key as an index, we'll
perform a linear search to find the command that is mapped to a key.
This removes the requirement to call init functions for the commands.

Insert mode now enters its own command loop, this might not be final
though.