Implement iterator for Gap_buffer
Before we were using an indexed iterator for the Gap_buffer, this
change introduces a more idiomatic iterator the gap buffer while
implementing the indexed iterator as a sort of adaptor. Better tests
need to be written for the gap buffer, but I've tried to catch as many
possible problems with the new iterator design.