Add command counting
Start work on providing a count to normal commands, will need to guard
against overflow and in the future display the current value. Could also
swtich to an emacs style count, where we prompt the user for a value.

I've only added `forward_char` command to support counting for testing.