Fix warning about stack usage
Visual Studio givens a warning (C6262) when allocating stack space to
restore the title for the console. We now will allocate on the heap
using std::string and restore the title through that.

I've also added a note about the potential for using GetConsoleTitle at
the start of the application rather than restoring the "original" title.